# [api] Optimise `TraceContextPropagator.Extract`