@stevejgordon 2024-08-07
@CodeBlanch, I am just bumping here to keep this open and see if we can solve the discussion on using `stackalloc` offline to move this along? I had hoped to attend the SIG to raise it, but the timing didn't work out again last night.